requirements.txt
appnope==0.1.0
argcomplete==1.9.2
asn1crypto==0.24.0
awscli==1.14.63
base58==0.2.4
bcrypt==3.1.4
bleach==2.1.3
boto==2.48.0
boto3==1.6.15
botocore==1.9.16
certifi==2018.1.18
cffi==1.11.5
cfn-flip==1.0.3
chardet==3.0.4
click==6.7
colorama==0.3.7
cryptography==2.2.1
decorator==4.2.1
docutils==0.14
durationpy==0.5
entrypoints==0.2.3
Fabric==1.14.0
Flask==0.12.2
future==0.16.0
hjson==3.0.1
html5lib==1.0.1
idna==2.6
ipykernel==4.8.2
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.2
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.10
jmespath==0.9.3
jsonpickle==0.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kappa==0.6.0
lambda-packages==0.19.0
MarkupSafe==1.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.1
numpy==1.14.2
pandas==0.22.0
pandocfilters==1.4.2
paramiko==2.4.1
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
placebo==0.8.1
prettytable==0.7.2
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyasn1==0.4.2
pycparser==2.18
Pygments==2.2.0
PyNaCl==1.2.1
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2018.3
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.13
scikit-learn==0.19.1
scipy==1.0.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sklearn-pandas==1.6.0
terminado==0.8.1
testpath==0.3.1
toml==0.9.4
tornado==5.0.1
tqdm==4.19.1
traitlets==4.3.2
troposphere==2.2.1
Unidecode==1.0.22
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12
widgetsnbextension==3.1.4
wsgi-request-logger==0.4.6
zappa==0.45.1
